skip assets precompile when assets not changed

+++ config/deploy.rb
@@ -38,14 +38,20 @@ namespace :deploy do
   end
 
   namespace :assets do
+
     desc "Local precompile assets and upload to server"
     task :precompile, roles: :app do
-      run_locally "RAILS_ENV=#{rails_env} #{rake} assets:clean && RAILS_ENV=#{rails_env} #{rake} assets:precompile"
-      run_locally "cd public && tar -jcf assets.tar.bz2 assets"
-      top.upload "public/assets.tar.bz2", "#{shared_path}", via: :scp
-      run "cd #{shared_path} && tar -jxf assets.tar.bz2 && rm assets.tar.bz2"
-      run_locally "rm public/assets.tar.bz2"
-      run_locally "#{rake} assets:clean"
+      from = source.next_revision(current_revision)
+      if capture("cd #{latest_release} && #{source.local.log(from)} vendor/assets/ lib/assets/ app/assets/ | wc -l").to_i > 0
+        run_locally "RAILS_ENV=#{rails_env} #{rake} assets:clean && RAILS_ENV=#{rails_env} #{rake} assets:precompile"
+        run_locally "cd public && tar -jcf assets.tar.bz2 assets"
+        top.upload "public/assets.tar.bz2", "#{shared_path}", via: :scp
+        run "cd #{shared_path} && tar -jxf assets.tar.bz2 && rm assets.tar.bz2"
+        run_locally "rm public/assets.tar.bz2"
+        run_locally "#{rake} assets:clean"
+      else
+        logger.info "Skipping asset precompilation because there were no asset changes"
+      end
     end
 
     desc "Symlink local precompile assets"
